Renamed some files. Moved one to impl.
