+++ src/batched/ode/KokkosBatched_ODE_SolverStatus.hpp
@@ -42,8 +42,8 @@
 //@HEADER
 */
 
-#ifndef __KOKKOSBATCHED_ODE_SOLVERENDSTATUS_HPP__
-#define __KOKKOSBATCHED_ODE_SOLVERENDSTATUS_HPP__
+#ifndef __KOKKOSBATCHED_ODE_SOLVERSTATUS_HPP__
+#define __KOKKOSBATCHED_ODE_SOLVERSTATUS_HPP__
 
 #include <ostream>
 namespace KokkosBatched {
@@ -56,20 +56,14 @@ enum class ODESolverStatus {
   NONFINITE_STATE
 };
 
-std::ostream& operator<<(std::ostream & os, ODESolverStatus  & status) 
-{
+std::ostream& operator<<(std::ostream& os, ODESolverStatus& status) {
   switch (status) {
-  case ODESolverStatus::SUCCESS:
-      os << "Success!";
-      break;
-  case ODESolverStatus::FAILED_TO_CONVERGE:
-      os << "Failed to Converge";
-      break;
-  case ODESolverStatus::MINIMUM_TIMESTEP_REACHED:
+    case ODESolverStatus::SUCCESS: os << "Success!"; break;
+    case ODESolverStatus::FAILED_TO_CONVERGE: os << "Failed to Converge"; break;
+    case ODESolverStatus::MINIMUM_TIMESTEP_REACHED:
       os << "Minimum Timestep Reached";
       break;
-  case ODESolverStatus::NONFINITE_STATE:
-      os << "Non-Finite State";
+    case ODESolverStatus::NONFINITE_STATE: os << "Non-Finite State";
   }
   return os;
 }
