+++ src/batched/ode/impl/KokkosBatched_ODE_RKSolve_Impl.hpp
@@ -50,11 +50,11 @@
 #include "Kokkos_MemoryTraits.hpp"
 #include "Kokkos_NumericTraits.hpp"
 
-#include <KokkosBatched_ODE_Args.h>
-#include <KokkosBatched_ODE_RungeKuttaTables.h>
-#include <KokkosBatched_ODE_SolverEndStatus.h>
+#include <KokkosBatched_ODE_Args.hpp>
+#include <KokkosBatched_ODE_RungeKuttaTables.hpp>
+#include <KokkosBatched_ODE_SolverStatus.hpp>
 
-#include <type_traits>
+//#include <type_traits>
 
 namespace KokkosBatched {
 namespace Experimental {
